+++ changelog.md
@@ -14,3 +14,4 @@
 
 - Se agrego la logica a carrito
 - Se cambio la carga de las imagenes y elementos a cargar con un arreglo predefinido
+- Se agrego la logica para la persistencia con localStorage para el carrito